refactoring: move offchain v2 related code into offchain_api_v2.py, and make it an option to setup miniwallet app
